Make the Bedstead Web page unambiguously XHTML.
It was syntactically XHTML already, but various things suggested that it
might be HTML instead.  Now it has a proper content-type in Apache, a
filename that Firefox can understand when reading from a local file, and
for good measure a <meta> tag.  That should keep everyone except possibly
Lynx users happy.