Add support for mosaic graphics.
These are generated without smoothing, and this is done
algorithmically, bypassing the usual character-rendering code.  The
characters are put in the private use area, in the locations where
zvbi uses for them, on the grounds that we have to put them somewhere.

Mosaic graphics characters are not yet unified with the block graphics
characters in the U+2580 range.